Increased sub sample to u16

+++ src/bench_thread_geom.zig
@@ -32,7 +32,7 @@ const F = buildconfig.F;
 const DEFAULT_OUT_DIR = "out/bench_stats_thread_geom";
 const DEFAULT_IMAGE_OUT_DIR = "out/bench_images_thread_geom";
 const DEFAULT_PIXELS_NUM = [2]u32{ 1600, 1000 };
-const DEFAULT_SUB_SAMPLE: u8 = 1;
+const DEFAULT_SUB_SAMPLE: u16 = 1;
 const DEFAULT_FOCAL_LENG: F = @floatCast(50.0e-3);
 const DEFAULT_PIXELS_SIZE = [2]F{
     @floatCast(5.3e-6),
